server/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
Werkzeug==2.3.8

# Медиа
Pillow==10.4.0
piexif==1.1.3

# HTTP
requests==2.31.0

# БД: PostgreSQL/PostGIS
psycopg2-binary==2.9.9

# Продакшен-сервер
gunicorn==21.2.0
